# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chore Club API",
    "description": "Multi-user, multi-household, multi-device chore tracker. RFC 7807 errors, `If-Match` optimistic concurrency on aggregate roots, `Idempotency-Key` on side-effecting hot-path POSTs, and a delta-sync protocol on `/sync`. Authoritative spec: `docs/api-spec.md`.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local docker-compose stack."
    },
    {
      "url": "https://api.choreclub.app",
      "description": "Production."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 access token issued by `/v1/auth/google` or `/v1/auth/apple`. 15-min TTL."
      }
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "description": "UUIDv7 string."
      },
      "IsoTimestamp": {
        "type": "string",
        "format": "date-time",
        "description": "ISO-8601 UTC, ms precision."
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 problem+json body. Returned for every error.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `household.not_a_member`."
          },
          "errors": {
            "description": "Optional structured payload (validation issues, conflict snapshot, etc.)."
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "emailVerified",
          "displayName",
          "avatar",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "emailVerified": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "avatar": {
            "type": "string",
            "maxLength": 16
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresAt",
          "user"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "AppleNonceResponse": {
        "type": "object",
        "properties": {
          "nonce": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "nonce",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "RefreshResponse": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresAt"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        }
      },
...
      "GoogleSignInBody": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "minLength": 20
          },
          "deviceLabel": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "idToken"
        ],
        "additionalProperties": false
      },
      "AppleSignInBody": {
        "type": "object",
        "properties": {
          "identityToken": {
            "type": "string",
            "minLength": 20
          },
          "nonce": {
            "type": "string",
            "minLength": 8
          },
          "deviceLabel": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "identityToken",
          "nonce"
        ],
        "additionalProperties": false
      },
      "RefreshBody": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 20
          }
        },
        "required": [
          "refreshToken"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/auth/apple/nonce": {
      "post": {
        "summary": "Issue a single-use nonce for the next Apple sign-in (5-min TTL).",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppleNonceResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [],
        "operationId": "postAuthAppleNonce"
      }
    },
    "/v1/auth/google": {
      "post": {
        "summary": "Sign in with a Google ID token. Returns a session + user profile.",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInBody"
              }
            }
          }
        },
        "security": [],
        "operationId": "postAuthGoogle"
      }
    },
    "/v1/auth/apple": {
      "post": {
        "summary": "Sign in with an Apple identity token + previously issued nonce.",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleSignInBody"
              }
            }
          }
        },
        "security": [],
        "operationId": "postAuthApple"
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "summary": "Rotate a refresh token. Replay → 401 + family revoked.",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          }
        },
        "security": [],
        "operationId": "postAuthRefresh"
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Revoke the current refresh-token family.",
        "tags": [
          "auth"
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "postAuthLogout"
      }
    },
    "/v1/auth/logout-all": {
      "post": {
        "summary": "Revoke every refresh-token family for the user.",
        "tags": [
          "auth"
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "postAuthLogoutAll"
      }
    },
    "/v1/auth/account": {
      "delete": {
        "summary": "Begin account deletion. Refuses if the user owns multi-member households.",
        "tags": [
          "auth"
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — stale `If-Match`, idempotency-key reuse, or domain conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "deleteAuthAccount"
      }
    },
...
  }
}